+++ api/api_pesanandetail/src/index.js
@@ -23,60 +23,41 @@ app.get('/api/readpesanandetail',(req,res)=>{
 })
 
 
-//create
-app.post('/api/createartikel', (req, res) => {
-    const gambarArtikel = req.body.gambar;
-    const deskripsiArtikel = req.body.deskripsi; 
-    const judulArtikel = req.body.judul_artikel; 
-  
-    const sqlQuery = "INSERT INTO artikel (gambar, deskripsi, judul_artikel) VALUES (?, ?, ?)";
-    db.query(sqlQuery,[gambarArtikel,deskripsiArtikel, judulArtikel ], (err,result)=>{
-        if(err) {
-            console.log(err);
-
-        }else{
-            res.send(result);
-            console.log(result)
-        }
-    })
-  });
-
-  //update
-  app.put('/api/updateartikel/:id', (req, res) => {
-    const articleId = req.params.id;
-    const gambarArtikel = req.body.gambar;
-    const deskripsiArtikel = req.body.deskripsi; 
-    const judulArtikel = req.body.judul_artikel; 
-    
-    const sqlQuery = "UPDATE artikel SET gambar=?, deskripsi=?, judul_artikel=? WHERE id_artikel=?";
-    db.query(sqlQuery, [gambarArtikel, deskripsiArtikel, judulArtikel, articleId], (err, result) => {
-      if (err) {
-        console.log(err);
-        res.status(500).send(err); // Send an error response to the client
-      } else {
-        res.send(result);
-        console.log(result);
-      }
-    });
-});
-
+ 
+ 
 
   //delete
 
   app.delete('/api/deletepesanandetail', (req, res) => {
-    const idPesananDetails  = req.body.id; // Fix: Corrected variable name
-    const sqlQuery = "DELETE FROM pesanan_details WHERE id = ?";
-  
-    db.query(sqlQuery, [idPesananDetails], (err, result) => {
-      if (err) {
-        console.error(err);
-        res.status(500).send("Error deleting detail pesanan");
-      } else {
-        console.log(result);
-        res.send(result);
-      }
+    const idPesananDetails = req.body.id;
+
+    // Periksa apakah detail pesanan dengan ID yang diminta ada dalam database sebelum melakukan delete
+    const checkQuery = "SELECT * FROM pesanan_details WHERE id = ?";
+    db.query(checkQuery, [idPesananDetails], (checkErr, checkResult) => {
+        if (checkErr) {
+            console.log(checkErr);
+            return res.status(500).send("Internal Server Error");
+        }
+
+        // Detail pesanan tidak ditemukan dalam database
+        if (checkResult.length === 0) {
+            return res.status(404).send("Detail pesanan not found");
+        }
+
+        // Lakukan delete jika detail pesanan ditemukan
+        const sqlQuery = "DELETE FROM pesanan_details WHERE id = ?";
+        db.query(sqlQuery, [idPesananDetails], (err, result) => {
+            if (err) {
+                console.error(err);
+                res.status(500).send("Error deleting detail pesanan");
+            } else {
+                console.log(result);
+                res.send(result);
+            }
+        });
     });
-  });
+});
+
